Implements AlphaBacktest
Create a GetAlphaBacktest method in SDK that creates a project, creates an algorithm that use historical alpha streams insights to trade, fetchs the backtest result and delete the project.
The backtest result is used to create the AlphaBacktest object where the data is organized as dictionaries or pandas.DataFrame to make visualization and manipulation easier.